requirements.txt
Pillow==2.4.0
requests==2.2.1
six==1.6.1